Update controller.md to reflect newly scaffolded code.
Tutorial still maintains the use of "this" as variable name for MainController struct. Bee v1.10.0 generates new projects using "c" as variable name for MainController struct. The use of "this" conflicts with the latest version of Bee and it is very likely for the tutorial to create confusion in people that are new to Beego.